Skylight and weather checking
Scrapping direct LOS checking due to excessive computational load. 

Panels now have a 120 degree swing (+60 to -60, relative to noon) over
the course of the day, with efficiency dropping off towards
sunrise/sunset due to oblique angles.

Rounding for storage done as late as possible on a per-tick basis -
rounding on a per-second basis would be slightly more accurate, but at a
cost of spiky generator output.